// console/src/api/generated/models/link-verification-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.24.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Request body for triggering link verification.
 * @export
 * @interface LinkVerificationRequest
 */
export interface LinkVerificationRequest {
    /**
     * Metadata name of the link group to verify when names is empty.
     * @type {string}
     * @memberof LinkVerificationRequest
     */
    'groupName'?: string;
    /**
     * Metadata names of links to verify. Takes precedence over groupName.
     * @type {Array<string>}
     * @memberof LinkVerificationRequest
     */
    'names'?: Array<string>;
}